feat: article CRUD auth

+++ src/router/api/article/article.controller.ts
@@ -81,12 +81,13 @@ export const getArticle = async (req: Request, res: Response) => {
   }
  */
 export const writeArticle = async (req: Request, res: Response) => {
-  const { title, writer, content } = req.body;
+  const { title, content } = req.body;
+  const writer = req.decoded.uid;
   const reg_date = moment(new Date()).format("YYYY-MM-DD HH:mm:ss");
 
   try {
     await Article.create({ title, writer, content, reg_date });
-    res.status(201).json({ success: 1 });
+    res.status(201).json({ success: true });
   } catch(err) {
     res.status(400).json({ error: err });
   }
@@ -116,32 +117,33 @@ export const modifyArticle = async (req: Request, res: Response) => {
 };
 
 /*
-  DELETE /api/article/delete/:id
+  DELETE /api/article/delete?id=?
  */
 export const deleteArticle = async (req: Request, res: Response) => {
-  const { id } = req.params;
+  const id: number = +req.query.id!;
+  const { uid } = req.decoded;
   let transaction = null;
 
   try {
     transaction = await sequelize.transaction();
+
+    const article: Article | null = await Article.findOne({ where: { article_no: id } });
+    if(uid !== article?.writer) throw "삭제할 권한이 없습니다.";
+
     await Reply.destroy({
       where: {
         article_no: id
       },
       transaction
     });
-    await Article.destroy({
-      where: {
-        article_no: id
-      },
-      transaction
-    });
+    await article?.destroy({ transaction });
+    
     transaction.commit();
-    res.status(204).json({ success: 1 });
+    res.status(204).send();
   } catch(err) {
     transaction!.rollback();
     res.status(400).json({
-      success: 0,
+      success: false,
       error: err
     });
   }
